AmountConverter in extra file and fixed ugly formatting
